move: enum 클래스 각 도메인별 enums 패키지로 이동
