+++ src/main/java/kr/hhplus/be/server/domain/balance/UserBalanceHistory.java
@@ -1,6 +1,7 @@
 package kr.hhplus.be.server.domain.balance;
 
 import jakarta.persistence.*;
+import kr.hhplus.be.server.domain.balance.enums.TransactionType;
 import kr.hhplus.be.server.global.entity.BaseEntity;
 import kr.hhplus.be.server.global.exception.ApiErrorCode;
 import kr.hhplus.be.server.global.exception.ApiException;
