+++ src/main/java/kr/hhplus/be/server/domain/balance/UserBalanceService.java
@@ -1,5 +1,6 @@
 package kr.hhplus.be.server.domain.balance;
 
+import kr.hhplus.be.server.domain.balance.enums.TransactionType;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
