+++ src/main/java/kr/hhplus/be/server/domain/coupon/CouponService.java
@@ -1,5 +1,6 @@
 package kr.hhplus.be.server.domain.coupon;
 
+import kr.hhplus.be.server.domain.coupon.enums.UserCouponStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
