+++ src/main/java/kr/hhplus/be/server/domain/coupon/UserCouponInfo.java
@@ -1,5 +1,8 @@
 package kr.hhplus.be.server.domain.coupon;
 
+import kr.hhplus.be.server.domain.coupon.enums.DiscountType;
+import kr.hhplus.be.server.domain.coupon.enums.UserCouponStatus;
+
 import java.time.LocalDateTime;
 
 public record UserCouponInfo(
