+++ src/main/java/kr/hhplus/be/server/domain/coupon/enums/CouponStatus.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.domain.coupon;
+package kr.hhplus.be.server.domain.coupon.enums;
 
 public enum CouponStatus {
     ACTIVE,
