+++ src/main/java/kr/hhplus/be/server/domain/coupon/enums/DiscountType.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.domain.coupon;
+package kr.hhplus.be.server.domain.coupon.enums;
 
 public enum DiscountType {
     PERCENTAGE, // 퍼센트 할인 (예: 10%)
