+++ src/main/java/kr/hhplus/be/server/domain/order/enums/OrderStatus.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.domain.order;
+package kr.hhplus.be.server.domain.order.enums;
 
 public enum OrderStatus {
     PENDING,
