+++ src/main/java/kr/hhplus/be/server/interfaces/coupon/dto/CouponIssueResponse.java
@@ -1,8 +1,8 @@
 package kr.hhplus.be.server.interfaces.coupon.dto;
 
-import kr.hhplus.be.server.domain.coupon.DiscountType;
+import kr.hhplus.be.server.domain.coupon.enums.DiscountType;
 import kr.hhplus.be.server.domain.coupon.UserCouponInfo;
-import kr.hhplus.be.server.domain.coupon.UserCouponStatus;
+import kr.hhplus.be.server.domain.coupon.enums.UserCouponStatus;
 
 import java.time.LocalDateTime;
