+++ src/test/java/kr/hhplus/be/server/domain/balance/UserBalanceServiceTest.java
@@ -1,5 +1,6 @@
 package kr.hhplus.be.server.domain.balance;
 
+import kr.hhplus.be.server.domain.balance.enums.TransactionType;
 import kr.hhplus.be.server.domain.user.User;
 import kr.hhplus.be.server.global.exception.ApiErrorCode;
 import org.junit.jupiter.api.BeforeEach;
